+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [64.1.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@64.1.3...@codecademy/gamut@64.1.4) (2025-06-17)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+### [64.1.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@64.1.2...@codecademy/gamut@64.1.3) (2025-06-16)
+
+**Note:** Version bump only for package @codecademy/gamut
+
 ### [64.1.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@64.1.1...@codecademy/gamut@64.1.2) (2025-06-12)
 
 **Note:** Version bump only for package @codecademy/gamut